move splitfuncs.py from HeppyResult to Loop
